move profiler to own file
